removed old files, modified game.rs
