+++ bdi_game/src/display/mod.rs
@@ -1,6 +1,5 @@
 pub mod actions;
-pub mod rendering;
+pub mod hexagon;
+pub mod renderable;
 pub mod resource_manager;
-pub mod sdl;
 pub mod structures;
-pub mod traits;
